requirements.txt
colorama==0.4.6
eth_account==0.13.5
imap_tools==1.10.0
loguru==0.7.3
playwright==1.50.0
primp==0.13.0
pyfiglet==1.0.2
pytz==2025.1
PyYAML==6.0.2
Requests==2.32.3
termcolor==2.5.0
urllib3==2.3.0
web3==7.8.0
tabulate==0.9.0
rich
prompt_toolkit